+++ jmicro.api/src/main/java/org/jmicro/api/net/IMessageHandler.java
@@ -23,7 +23,7 @@ package org.jmicro.api.net;
  */
 public interface IMessageHandler {
 
-	Short type();
+	Byte type();
 	
 	void onMessage(ISession session,Message messge);
 }
